size

+++ 0x00-hello_world/6-size.c
@@ -5,10 +5,10 @@
  */
 int main(void)
 {
-	printf("size of a char: %d byte(s)\n", sizeof(charType));
-	printf("size of an int: %d byte(s)\n", sizeof(intType));
-	printf("size of a long int: %d byte(s)\n", sizeof(longType));
-	printf("size of a long long int: %d byte(s)\n", sizeof(longlongType));
-	printf("size of a float: %d byte(s)\n", sizeof(floatType));
+	printf("size of a char: %d byte(s)\n", (int) sizeof(charType));
+	printf("size of an int: %d byte(s)\n", (int) sizeof(intType));
+	printf("size of a long int: %d byte(s)\n", (int) sizeof(longType));
+	printf("size of a long long int: %d byte(s)\n", (int) sizeof(longlongType));
+	printf("size of a float: %d byte(s)\n", (int) sizeof(floatType));
 	return (0);
 }
